+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Added command **Generate messages** (`dftemplate.generateMessages`), also available as a code action on the `QRC` directive, which generates missing messages selected from a prompt.
 - Refactoring: switch among message text aliases if more than one is available.
 - Added command **Order messages** (`dftemplate.orderMessages`), also available as a code action, which sorts messages by id.
+- Added command **Generate global variables** (`dftemplate.generateGlobalVariables`), also available as a code action on the `QBN` directive, which generates missing global variables selected from a prompt.
 
 ### Improved